skedder/requirements.txt
arrow==0.13.0
beautifulsoup4==4.7.1
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.8
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.8.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.7.3
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.4.0
